+++ OpenNos.GameObject/Helpers/EventHelper.cs
@@ -333,7 +333,7 @@ namespace OpenNos.GameObject.Helpers
                                                     rare = (sbyte) ServerManager.Instance.RandomNumber(-2, 7);
                                                 }
                                                 //TODO add random rarity for some object
-                                                sess.Character.GiftAdd(gift.VNum, gift.Amount, rare, gift.Design);
+                                                sess.Character.GiftAdd(gift.VNum, gift.Amount, (byte)rare, gift.Design);
                                             }
                                         }
                                     }
